Forslag til forbedringer. (#20)
