+++ .eslintrc
@@ -34,7 +34,7 @@
     "react-hooks/exhaustive-deps": "error",
     "react/jsx-uses-vars": ["warn"],      // avoid no-unused-vars for imports used as react components
     "react/jsx-uses-react": ["warn"],     // avoid no-unused-vars for importing React
-    "react/jsx-no-bind": ["error", {"allowArrowFunctions": true}],
+    "react/jsx-no-bind": ["error", {"allowArrowFunctions": true}], // See https://stackoverflow.com/a/53177820
     "react/react-in-jsx-scope": ["warn"], // make sure React is in scope when writing jsx
     "no-var": ["error"],
     "jsx-a11y/label-has-for": [ "error", {
