+++ src/utils/renderMicrobit.js
@@ -1,6 +1,23 @@
 const microbitIframeId = 'makecoderenderer'; 
 const getMicrobitSnippets = () => Array.from(document.getElementsByClassName('microbit'));
 
+const renderSpinner = () => {
+  const pres = [...document.getElementsByTagName('pre')];
+  for (let pre of pres) {
+    if ([...pre.childNodes][0].className === 'python') return;
+    pre.style.display = 'none';
+    let img = document.createElement('img');
+    img.id = 'spinner';
+    img.src = require('../assets/graphics/spinner.gif');
+    img.alt = 'Spinner';
+    img.width = '50';
+    img.height = '50';
+    img.style.maxWidth = '100%';
+    img.style.display = 'block';
+    img.style.margin = '0 auto 15px';
+    pre.parentElement.insertBefore(img, pre);
+  }
+};
 
 /**
  * Creates an iframe that is being used to render the code
@@ -110,6 +127,7 @@ const processIframeMessage = e => {
 
 export const renderMicrobit = language => {
   if (getMicrobitSnippets().length > 0) {
+    renderSpinner();
     createIframe(language);
   }
 };
